Disable check function
Check function was causing a change indata type(?) and maybe something
else. Could not do a < operator between a numpy array and a str. And if
I fixed that it threw a warning about modifying a slice. Something about
chained commands. Anyway, to hell with it. I'm reverting back to
something that works so I can learn other stuff.